Don't claim 3rd-party repos are part of the OS on Debian derivatives
This improves https://invent.kde.org/plasma/discover/-/merge_requests/474 on Debian
and its derivatives (Ubuntu, PureOS, Pop_OS, ...) by only showing the distribution
as the origin if it actually is the origin, and showing the raw repository name in
any other case.

While the raw repo ID string is not pretty at all, and PackageKit still lacks an API
to get that pretty name (which APT actually knows...) at least this change is more
honest than claiming things are part of the OS that actually aren't.