Flatpak: Fix spacing in permissions view
ColumnLayout should always specify one of the standard units for spacing.
In this case we set it to 0, because most of it is a list of delegates,
but header and footer now have their own margins on the list's side.

(cherry-picked only what's applicable from commit
b54c72119197bf05e16db05577c4b7265521068f)