rpm-ostree/notifier: Setup a watcher to trigger reboot check
To make sure that the notifier rechecks if a reboot is needed, we need
to watch for changes made to ostree deployments.

To avoid triggerring the check mutiple times or while the rpm-ostree
transaction is still running, we add a 10 second delay.

(cherry picked from commit 0534a8a7e3b0a3dff7489607de9c501adf250ca1)