ApplicationResourceButton: attribute the left/right padding
Currently we expose the label width via implicitTitleTextWidth to the
caller. Who then hard-codes the left/right padding, trying to calculate
the actual minimal width.

Regardless - we shouldn't be doing that, since it also implies that it
should know about the exact layout. Just pass the actual (minimal) width
instead.

Setting Layout.minimumWidth: metric.width for Kirigami.Heading does not
quite work. So instead just use a custom property.

Note: the stackedlayout calculation is off regardless of this patch.
That will be fixed with follow-up commit.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>