ApplicationPage: Allow main app info column to grow with window
Right now the width of the main column on this page has a fixed maximum
size. When the window is reeeeeally wide because the user maximized it,
the result can look a bit silly, with excessive whitespace and not
enough room for the buttons to have a horizontal layout.

To fix this, let's allow the maximum size to expand with the page, while
still preserving some whitespace on the sides to avoid letting the
layout become silly-looking in the other direction.