notifier: Don't show updates notification if Discover is running
At times, we'll get new updates when Discover-the-app is running. It's
not really useful to show the notification then as it's already
displayed on the UI.

It also saves us from notifying about changes when updates are getting
installed, this is sometimes a problem.


(cherry picked from commit 0f98292f1e1245312734dcb13107671764378dcc)