rpm-ostree: Improve handling of externally started transactions
While Discover (GUI), the notifier and the updater share the same
rpm-ostree backend code, a transaction started by one of the three will
currently appear as "externally started" to the other ones, as if it had
been manually started by a user via the command line for example.

Improve how we detect and handle transactions that the current backend
did not start and make sure that we check for in-progress transactions
before looking for updates.