fwupd: don't refresh directory-type remotes on setup
File- and directory-type remotes cannot be refreshed, attempting
to do so will trigger an error "no metadata URIs for (remote)"
(because, internally, fwupd does not set the metadata_uri
property for file- or directory-type remotes). We already avoid
trying to refresh file-type remotes on setup, but we need to do
the same for directory-type remotes. This fixes the error
appearing with fwupd 1.9.1 (which ships with a directory-type
remote, vendor-directory, enabled by default).

Signed-off-by: Adam Williamson <awilliam@redhat.com>

BUG: 462931