ApplicationPage: make heading/short description text selectable
The main text is already selectable, and the heading/short description
text is used to display the package path for locally-downloaded
packages, which make sense to be able to select. So for those reasons,
make it also selectable.

This entails porting away from Kirigami.Heading and using
Kirigami.SelectableLabel, which means we need to do a small amount of
work to match the visual styling and accessibility metadata that were
previously provided by Kirigami.Heading.

BUG: 470297
FIXED-IN: 5.27.6