pk: Notify of a new distro upgrade from AppStream
PackageKitNotifier already had the logic to notify of a distro
upgrade from
    PackageKit::Daemon::getDistroUpgrades().
However, not all pk backends implement that that.

In the PackageKitBackend we already implemented upgrading using
versions from AppStream, so check the same metadata in the notifier.

I have no evidence that the old code path was actually being used
by anyone: in facts the notification was misconfigured and thus
couldn't be shown, on modern KF versions at the least.
Nonetheless the old logic remains intact while this is implemented
in parallel.

Contrary to the old logic that invokes
    PackageKit::Daemon::upgradeSystem(version)
immediatly without showing any progress indicator, this notification
simply launches the discover UI where the already present upgrade
logic will guide the user through an informed distro upgrade.