Don't cap content width on Application page
It leads to the content being not vertically aligned with anything, or
aligned with random page elements based on the window size. This was
done in the past to prevent the content are from becoming too wide,
based on the logic that wider text is harder to parse. But this is not
worth it; just don't do it at all.