Rating: Rework the component for better UX and a11y
This is a library-grade component which I would be happy to upstream
into Kirigami and reuse in other places, such as KNewStuff and Elisa.

It supports keyboard interaction, RTL layout mirroring, optional
half-star display mode, and different editing policies.

Its interaction area now spans entire control (and slightly beyond --
across control's paddings, for convenience), and allows for press and
drag editing.

A minor improvement opportunity would be to support fractional values
beyond half-star precision, but it is not currently possible with a set
of 3 fixed icons.

One Kirigami-specific caveat, mouse stealing prevention had to be
enabled in order to make drag actually work in OverlaySheet.