Rework layouts and base component of application delegate cards
Sad but true, Kirigami.AbstractCard is an over-complicated engineering
effort, out of which we only really need its private background
component. When it comes to content sizing management, it lacks
required flexibility, which is why Discover components ended up abusing
it with a fake content height and inner layout's vertical centering.

Until that is fixed upstream one way or another, the best we could do
downstream is to take AbstractCard as a base (again, just for the sake
of its background), and redo its content management. Which isn't too
hard or complicated, so IMHO it's justified.

Application Delegate types got rid of hacks, and got a couple of
improvements to their layouts, such as consistent margins, standard
spacing on Row/ColumnLayout items, and icon animations disabled.