LabelBackground: Fix glitched graphics at near-zero values
When this custom progress bar component draws a rectangle that has less
height than the sum of its left & right corner radii, the result is a
thin vertical rectangle sticking outside of background.

The fix is to hide progress bar groove when it's less that double radius.
Which is practically what PlasmaComponents.3/ProgressBar is doing.