LabelBackground: Implement better, Breeze-like styling
Breeze style uses a combination of Window/textColor and highlight color
with varying alpha values to draw its QProgressBar with a nice border
and a fitting background color.

Had to use hard-coded opacity component, because in QtQuick Qt.alpha
function only overrides base color's alpha, but Breeze uses
multiplication to obtain progressively smaller fractions. In the end of
the day, it doesn't terribly matter, as long as it looks fine.