ApplicationPage: Fix random gap at the top, above header
This is a rather common caveat of layout types, and it's easy to fall
for, especially because it can go unnoticed for a long time.

If
- a layout is stretched to be larger than the combined implicit size of
  its children items (plus margins), and
- none of the items are set to fill the layout
  (fillWidth/fillHeight depending on what is the 'primary' direction
  for a particular layout type),
then said layout would use the additional space to distribute its items,
including putting semi-random extra maring at the beginning and the end.
It even disregards items' alignment like AlignTop or AlignLeft.

The solution here is simple: don't stretch the layout to fill Flickable
content item, but let the Flickable implicitly work the other way
around: adapt contentHeight to the actual content's height.