ApplicationPage: Fix gap under carousel caused by invisible InlineMessages
This required some trickstery, but there you have it for not managing
InlineMessage's padding internally. The biggest issue here, of course
it the inability of Repeater to provide a declarative API to track all
of its delegates in one expression; and the uselessness of its
visibleChildren property, since Loaders can't set them to be invisible,
because implicit propagation of visibility would break core logic of
InlineMessage component.