pk: Set cache-age for distro-upgrade transaction
The behaviour of different backends with respect to cache expiry is
not unique with the default cache-age setting.

In Fedora by default the cache is never expired. Instead, a patch
downstream is implemented to force refresh the cache. However, that
leaves untouched the cache of the non-current distro releases,
as used by UpgradeSystem.
It is thus important that we explicitly specify a cache-age hint
to unequivocally expire the cache after a while.