carousel: Port prev/next buttons to KirigamiComponents.FloatingButton
They are now filling the whole height of their parent view, they don't
accept any focus to avoid UX issues, and generally have a fresh visual
style.

As an experiment there is an adapter component that works in two modes:
by relying on ListView's currentIndex and count, and by tracking
ListView's atXBeginning/atXEnd (temporarily still needed for the old
inline view).

Might not look or work the best with the old dynamically sized popup
implementation, but this is gonna change in the following patches.