Improve UI of "Offline Updates" setting
We currently refer to this feature using its technical name, but this
can be confusing for users since the common definition of "offline" is
"without internet access" which is not the meaning it has in this
context. Instead, be more descriptive about what the UX will look like
with plain language, and use radio buttons to make both possible
options clear to the user.