Show sound system access permission
Back when designing this feature, we intentionally didn't show the sound
system access permission, reasoning that it was so common that it would
seem spammy.

However this turned out not to be the case, and apps are indeed
generally judicious with it. In addition, not showing it causes apps
that *only* have sound system permission to show no permissions at all,
which is wrong. In addition it creates an inconsistency with what's
shown in the Flatpak KCM.

That caused more problems than it solved, so let's just show when apps
have this access.