Only preload a couple of screens worth of comments
Discover used to preload every comment in order to have the scrollbar
behaving as good as possible with variable height items.
However some applications like Firefox have thousands of comments
which will freeze firefox for minutes when scrolling quickly,
eat a lot of memory and will make discover even slow to quit
in order to delete the thousands of items it created.

Cache only a couple of screens worth of comments, so that applications with few comments
behave well and applications with thousands remain efficient

BUG:474944