Better sidebar behavior on search
Right now when a search is active is possible to drill down in categories to refine the search, but is not possible to know what ones will yeld any result (and usually mostly won't) so makes this drilling down not much useful.

Moreover, by pressing back, the search is unexpectedly reset forcing the user to search again.

Make easier and much more intuitive navigate search results and narrow
them down, or broaden them until what has been looking for is found

* disable categories which don't have any results in
* pressing "back" won't reset search terms

![Screencast_20231020_100534-1](/uploads/03c74cd296f652c4d1d852389bd851ab/Screencast_20231020_100534-1.webm)