Small redesign of ApplicationPage
Changes:
- Use buttons for things that perform actions within the app
- Use links for things that open external URLs
- Unify and improve spacing between logical groups
- Remove redundant app name in some headings

It's sort of unnecessarily huge sitting on the page, and there's plenty
of room for it up on the toolbar. This saves a bit of space on the page
and facilitates turning the other buttons on this row into underlined
links to make it clear that they will open a page in your browser.

BUG: 475848
FIXED-IN: 6.2