Fix PaginateModel
The logic when to do a reset and when to do an insertRows was a bit
fragile, and it happened that in the beginning when the model was empty
the condition went into inserting rows, but after this was done the same
condition wasn't true anymore and and endResetModel() was called
instead.
use a member bool to track whether we are performing a model reset.