Fix network cache size for first run
QNetworkDiskCache will take care of creating a directory if it doesn't
exist. QStorageInfo on a directory that doesn't exist will return that
it has 0 bytesTotal.

As well as being potentially inefficient for the second run, it has been
reported that a QNetworkDiskCache of 0 causes out of bounds memory
writes.

CCBUG: 464517