Relevance sorting heuristics for search
Search also in appstreamId(): weird but only way to make gimp come up
when searching for gimp, there is also appstream searchTokens but they
are abused and lead to a lot of garbage results.

Do a new sortRole default when there is a search. This relevance role
combines:

* rating
* a measure based on The Levenshtein distance between search string and
  app name
* an exact match with the name or a sub-word of the name

It makes sortings much more "expected" than any of the other sortings.

BUG: 399502
BUG: 475844
BUG: 469026
FIXED-IN: 6.0

![image](/uploads/05fae0a2b90bebb70f3aa6da78e04c4e/image.png)