DiscoverWindow: Clean up currentTopLevel change handler
There is no third argument to PageRow::push() method, and API consumers
are not required to provide second either if it's just an empty object.
I couldn't even imagine what could that boolean result of status check
mean, but I suppose it might be a thing of the past.

Also, `window.` qualifier isn't needed, since we are literally in
top-level window's context.