Rename `busy` properties to be more conventional
Boolean properties are often prefixed with `is` in their getter but not
in their name. And sending a new value as a signal argument does not
add anything useful except requiring QML developers to remember to put
an explicit argument to change signal handlers (or else we get a
deprecation warning).