notifier: quit if notification frequency is never
We get a lot of complaints about not being able to disable the
update notifier, with major downstream distributions even taking the
drastic measure of making a split plasma-discover-notifier package.

Introduces a quick fix by just quitting if the notification frequency
is set to never. The downside is that updates to the notification frequency
will only take effect after reboot.

BUG: 413053
FIXED-IN: 6.0