Home page: span and wrap title labels
Otherwise in some languages, the label object will overflow the view
when it's very narrow, as in widescreen mode when clicking on one of the
apps on the page.

To prevent this, make the labels fill width and wrap.

BUG: 476898
FIXED-IN: 6.0


(cherry picked from commit 7b3771cb08ec4290a902489e9fd7f8e36d2d2e62)