Improve time to a ready startup page
Do not wait for the file to be downloaded to provide something that can
be shown on the main UI if we still have the featured data from previous
runs.
If this file was out of date, it probably wouldn't make a big difference
and it gives the user the peace that Discover is ready to use rather
than a loading screen that you don't know what has to offer.


(cherry picked from commit 1c78b31691ce7243f6e6b88f57a5903e0322f175)