Do not use the currentSubMenu to reset the view
This triggers for a moment the launch of the Browsing view (i.e. the
default one) at startup. This would load the view even though Discover
was started on updates mode (or any of the other).
The Browsing view is especially heavy because it fetches all apps to
create its recommendation views.