Flatpak: Rewrite racy signal propagation into a single property expression
The logic here was broken on multiple levels, and was highly dependent
on a particular order of change signals propagation. The mix of
reactive, imperative and declarative pieces turns this into an
interesting "what if I were a compiler" mind puzzle. Luckily, we can
avoid all that hustle with a simple iterative method, which still binds
to changes of the versionCompare property of all the action objects
accessed in its callback.