packagekitbackend: de-thread the appstream loading hotfix
the threading causes a race condition in module loading where GIO
(appstream,packagekit) and libsoup (flatpak) try to load modules at the
same time and get stuck on each other. disable threading to temporarily
fix this for 6.0

needs addressing in 6.1 in a more reasonable way. perhaps move all
plugin loading into a threadpool with size 1? or run the first appstream
loading on the gui thread? or have appstream provide an init function?

CCBUG: 481993