Fix up code formatting
Also, an important note about previous patch
fb6b7426f909e41d1dcb48533319d567890da010:

The lambda inside QtConcurrent::run is NOT a coroutine. It is a regular
lambda which can safely accept captures. The result of run() call is a
QFuture which an enclosing coroutine awaits for, and it has nothing to
do with how coroutine captures are destroyed at their first suspension.
So, whatever bug it was attempting to fix, it should have not changed
anything in the end; just a slightly more verbose way of writing the
same thing.