packagekitresource: do not load dependencies on construction
there is a limit on how many transaction we can have, don't exhaust them
by querying dependencies of updates when we'll possibly not need the
information at all.

this restores the previous behavior of only requesting dependency
information when sizeDescription is called. to facilitate this change
the dependencies class now has a getter on whether dependencies are
loaded

BUG: 464467