AppStreamIntegration: Add hidden option to enable updates to *development* versions
We currently have the `AllowPreRelease` hidden option that enables
updating to AppStream::Kind "Developement" releases which is used for
Fedora pre-releases such as Fedora 40 Beta.

Add an hidden option to enable updates to *development* versions (i.e
Rawhide for Fedora), which are AppStream::Snapshot.

Example `~/.config/discoverrc` config:

```
[DistroUpgrade]
AllowPreRelease=true
AllowDevelopmentRelease=true
```

This will be used in the rpm-ostree backend.