kns: drop provider part of the url
supposedly in 6.7 the parsing changed a bit and now certain providerIds
parse as path rather than host. the host component extraction was super
questionable anyway because it means you can't get to the providerId
from the url, rendering the provider part entirely pointless.

fortunately the provider part wasn't used anyway so we can simplify the
url and not include the provider at all

BUG: 486330