category: assert names aren't duplicated
this is an implicit requirement of our sorting logic that can break when
categories are non-uniquely translated

specifically it makes no sense to have ambiguous category names on the
same "level", the user isn't able to tell them apart (in particular also
in a a11y context -- e.g. screenreaders). so it makes no sense at all

guard against this with a simplistic test that loads all categories in
all known languages

CCBUG: 481303