packagekit: return empty license array when unknown
this was previously made non-empty so it looks like a valid license set.
since then we've grown the ability to actually show unknown licenses as
unknown in the UI, but that requires the license array to actually be
empty. so, make it empty again (:

BUG: 486037


(cherry picked from commit 5246ef454c5be0b16e3df7e232e4d3b3984435d0)