knewstuff: pick greatest linkId for installation
the implicit default of the kns api is '1' which would be the oldest id,
the newest id does kinda make more sense though. so sort by greater and
pass the greatest value into kns.

this is very much not ideal because links can be for anything so we may
be trading one set of wrong results for another set of wrong results
here if some store entries don't have the latest download link as the
latest id. there isn't really a good choice for us to make here since we
lack information. from OCS->attica->kns we don't get version data
unfortunately

BUG: 482534