flatpak: handle end_of_lifed_with_rebase
when end_of_lifed_with_rebase is called we need to render a decision on
whether to move to the replacement for the EOL flatpak or not. we do
this in two ways

1) runtime/ is always auto-transitioned because we consider this
implementation details that dont need surfacing to the user
2) everything else (i.e. app/) surfaces a proceed request via the
existing job interface

since we are dealing with a transaction thread the implementation uses a
bog standard waitcondition to implement the blocking aspect of this
(i.e. the callback doesn't return until the user made a proceed decision
causing the wait condition to get awoken)

BUG: 485886