flatpak: run transactions on a single thread
because of the potential need for authentication during the execution we
cannot just run N concurrent transactions. that works by default with
how flatpak polkit is set up, but if distros (i.e. suse) tweak those
settings to require auth_* then we are in a world of trouble. the hope
here is that by running everything on a single thread we'll be
serialized enough to allow for correct authentication serialization as
well

BUG: 466559