ApplicationPage: Do not use Item.visible to calculate the visibility of the parent
Having the parent visible=false will make all children visible:false so
it becomes a silly exercise.

Instead, use a property made for this that doesn't have other
connotations.

It kind of feels like we are workarounding a bug in the layout class but
this is needs addressing now.

BUG: 489745


(cherry picked from commit 101c01494ede10659740827c63dac874bf18f121)

Co-authored-by: Aleix Pol <aleixpol@kde.org>