Only chop off leading @ from tokenized license text if needed
Apparently at some point in the past, tokenized non-SPDX license labels
got prefixed with "@". As a result, the code unconditionally chopped
off the first character.

At some point, tokenizing stopped always doing that, causing some
license labels to be shown with their first letter missing.

I don't know if the prepending might come back in the future, so to be
safe, let's only conditionally remove the first character if it's "@".

BUG: 491172
FIXED-IN: 6.1.4