ApplicationPage: Be clearer about how we present and explain licenses
Right now Discover considers any non-free license to be proprietary. But
this is inaccurate; there are degrees of freedom, and a non-free license
is not necessarily proprietary.

This commit makes use of information already present to display a more
nuanced set of explanations based on the level of freedom that the
license offers:
- Free
- Non-Free
- Proprietary
- Unknown

BUG: 491151
FIXED-IN: 6.2.0