appstreamutils: suspend componentsByCategoriesTask for a bit
this should improve UI responsiveness while loading all apps, at the
cost of loading the apps slower.
the problem here is that as we basicaly enumerate all components known
to appstream we have a substantial amount of work to run through. this
takes a noticable amount of time even on fast systems. during this time
the UI is effectively frozen.

to resolve this problem the function is now a coroutine and suspends
itself for a bit of time so the eventloop has time to process inputs and
redraw itself.