Help people write better reviews
A lot of folks seem to have an instinct to write reviews saying things
like "it won't launch" and "it crashes a lot" which are not super useful
long term, as issues like those get fixed over time but the reviews
stick around forever.

In the proprietary software world, this kind of shouting into the void
may be the best you can do, but in our world, developers want to hear
about problems and respond directly to bug reports.

Accordingly, if an app isn't proprietary, help guide the user in a
productive direction if their review text matches certain keywords that
indicate they might be about to write a low-quality review.