ReviewDialog: use Dialog, not PromptDialog
We're not making use of any of PromptDialog's fancy features because we
have a custom contentItem to override all of it. Thus, it makes more
sense to use a basic Dialog, which is internally simpler and provides a
close button by default. This also gets rid of a ton of QML warnings.