ApplicationPage: display info for empty license
The recently added button that displays a license type explanation is
only shown if a license is set. If the list of licenses is empty, there
is a dedicated label displaying the license as "unknown", but this
dedicated label does not have an info button, as these are only added in
the repeater that iterates through actually existing license
information. The label does also not adjust its color.

This change adds the button for the missing license case, moves the
tooptip text into a property of the parent to reduce duplication, sets
the appropriate color for the license label, and adds translation
context.