ApplicationPage: fix unknown license text w/url
The license type explainer text for an unknown license where a project
url is known uses the <link> KUIT tag, but does not use xi18n, so the
link is not actually shown. In addition, the single-quoted placeholder
is not actually closed.

This change switches this text to xi18n and adds the closing single
quote.