AppStreamUtils: consider "unknown" licenses unknown
If a license is reported as "unknown" (rather than empty), the license
type function passes this through to the final free/non-free check,
where unknown licenses are always non-free. Discover has a special
category and explanations in the UI for unknown licenses though, and
translated string for this category, so it seems more appropriate to use
that rather than inform users about the non-free nature of what may well
be free software.

This change extends the cases where a license has unknown type to those
where the license is explicitly "unknown" (case-insensitively).