DiscoverWindow: change fetching updates label
The sidebar item for updates in Discover changes to "Fetching updates…"
when it is checking for updates. This has several issues: it uses
sentence case when all other sidebar entries use title case; it claims
to be fetching updates when it is actually not necessarily fetching the
updates themselves, but potentially only the information on which
updates are available; it looks like an action that requires further
user input as it ends in an ellipsis, and it mixes the category
("Updates") and the status ("Fetching") in a different way than the
final result, where the state "17 [Updates]" is placed in parentheses
after the category.

This change replaces the label with "Updates (Fetching…)", making case
consistent, avoiding the ambiguity from having a final ellipsis,
making it slightly more apparent that it may not be the actual
updates that are being fetched, but possibly only the metadata, and
making the progress status consistent with the result - the
"checking" turns into the number of updates, and vice versa.