Prompt for reboot after update again
In https://invent.kde.org/plasma/discover/-/merge_requests/899, this was
changed to show the logout screen with all options. However that doesn't
match the UI text of the button that takes you there.

Until we can offer a "reboot or shut down" prompt style and change the
UI here appropriately, return to the old style where the prompt matches
the button.