DiscoverWindow.qml: Port message OverlaySheet to Dialog and make it friendlier
Currently the passiveMessage dialog in DiscoverWindow is a bit too in-your-face with the errors, causing sadness and confusion with users.
It also used Kirigami.OverlaySheet even it is more of a dialog.

This MR makes it use Kirigami.Dialog features and hides the technical details behind a button, plus gives it a friendlier message.


BUG: 468657
BUG: 482551
FIXED-IN: 6.2.0