backends: switch API for categories
Instead of have resources describe its categories, just provide a
function to say whether a category name is provided.

This saves us for constructing thousands of vectors of strings for
categories every time we check whether to filter for a category.

Overall, it results in better performance and less CPU wastage.