ApplicationPage: Fix visibility of the reviews component
Do not make the visibility of a parent item depend on its children's as
if it affects them.
Also address the logic, it was weird. We need to show it if we are
fetching, if we fetched and didn't find anything or if we are trying to
show an error.

BUG: 495597