TransactionProgressIndicator: improve text readability
Previously it would alternate between the text color and highlighted
text color, but this didn't work since the highlighted text color is
only guaranteed to be readable on top of something using the highlight
color, and that couldn't be guaranteed here.

By contrast, the text color always works here, especially with a small
tweak to the colored area to desaturate it a tiny bit.

BUG: 492656
FIXED-IN: 6.2.4