DiscoverDrawer: impove sidebar keyboard navigation
The global drawer in Discover currently has inconsistent and partially
undiscoverable keyboard navigation. Arrow keys work for the parts of the
sidebar that represent categories, and they are reachable through the
tab key. But the actions that are fixed (including Home, Settings, etc.)
cannot be reached this way, only through accelerator shortcuts. While
this makes them still keyboard-accessible, they are invisible to users
relying on screen readers as they are outside the tab focus chain. And
even for sighted users, the inconsistency impairs usability. Moreover,
while it's possible to reach the search field using the arrow keys (by
moving past the end), it's not possible to escape the search field with
the arrow keys. Contrast with e.g. System Settings, where pressing down-
arrow from the search bar moves focus to the other sidebar entries below
in a seamless manner.

This change makes the manual ActionListItems part of the focus chain,
which automatically makes the Kirigami arrow navigation work. Arrow-
movement between the search field and actions below is hardcoded. A bit
of apparently dead code that blocked arrow navigation past the Updates
section (presumably the end of the list in a previous iteration of
Discover) was removed as well, to allow full arrow navigation within
the sidebar.